# Optimizing Stacked Borrows (part 1?): Cache locations of Tags in a Borrow Stack