@RalfJung 2021-12-16
Thanks a lot for this PR! Those are some impressive wins given the conceptually simple changes. :)

> I'm addressing the second loop by keeping track of the region of stack where there could be items granting Permission::Unique. This optimization is incredibly effective because Read access tends to dominate and many trips through this code path now skip the loop entirely.

Presumably this also has basically no memory cost, so I wonder how much just doing this would help?

> But as far as I can tell there are never enough Untagged items to be significant.

So could we save some memory by removing this part?

> These optimizations result in a 30% improvement in the the serde1 benchmark and an 85% (~6x) improvement in the mse benchmark. Which is pretty great because I didn't even develop these changes by running against those programs.

Note that I have no idea how suited these benchmarks are. It might be good to do at least superficial benchmarking of the code in `tests/run-pass/stacked-borrows` to avoid surprises for more "normal" code.

> But as far as I can tell from some very casual instrumenting, the root cause of the stacked borrows memory usage is that borrow stacks grow over time without bound and a RangeMap<Stack> will keep duplicating Stacks. The Clone+PartialEq activity that this generates eventually dominates as stacks grow, because that's O(n) where the HashMap optimization here gets darn close to O(1) for all other operations. I think it may be possible to ease this by nesting stacks, something like parent: Option<Rc<Stack>>, to avoid the expensive Clone operation until we actually need to modify the older elements.

That was one of my hypotheses for the main culprit. The other one is that even if RangeMap is fitting to the access pattern (so we don't keep splitting and merging subranges), just the fact that the stacks keep growing will be expensive in the long run -- even if we have perfect sharing, that's still quadratic since we keep iterating a larger and larger vector.