@saethlin 2021-12-17
> Thanks a lot for this PR! Those are some impressive wins given the conceptually simple changes. :)

Hah! I was afraid these would look prohibitively complex.

> So could we save some memory by removing this part?

We could, if we always tag raw pointers. Without `-Zmiri-tag-raw-pointers`, this data structure returns indexes in `find_granting` and we need the right index to invalidate the right part of the stack. Or at least that's what I've been able to glean so far.

As a sidenote: This is definitely the reason `-Zmiri-tag-raw-pointers` is faster with this PR. Without it, there are still linear searches of the untagged stack in `find_granting`, and the number of Untagged borrows is up to about half that of the Tagged ones in large stacks.

> Note that I have no idea how suited these benchmarks are.

Yeah. I'll cook up a little script to run through some more reasonable benchmarking code with the various SB options.

> just the fact that the stacks keep growing will be expensive in the long run -- even if we have perfect sharing, that's still quadratic since we keep iterating a larger and larger vector.

I suspect all iteration can be efficiently amortized out. If the parent stack from a split is stored in an `Rc`/`Arc` we can use `ptr_eq` to do a constant-time parent comparison. But there also other questions about that implementation turns out on a real workload (I haven't even started on it yet).