@RalfJung 2021-12-19
> We could, if we always tag raw pointers. Without -Zmiri-tag-raw-pointers, this data structure returns indexes in find_granting and we need the right index to invalidate the right part of the stack. Or at least that's what I've been able to glean so far.

Well, we could always use a linear scan instead, right? That will only slow down raw ptr accesses, and save some memory. This might or might not be a reasonable trade-off, not sure.

> Without it, there are still linear searches of the untagged stack in find_granting, and the number of Untagged borrows is up to about half that of the Tagged ones in large stacks.

That is odd, since the search should stop at the first untagged item it finds -- except when it is looking for a writeable item, then it will skip all the read-only untagged items.