@saethlin 2021-12-20
> That is odd, since the search should stop at the first untagged item it finds

You're right. I don't know what I was looking at before, but this is a good insight. Nearly every access of an Untagged item wants the topmost Untagged item (easily demonstrated with some `println!` and `cargo miri run | counts`. Removing the Vec of untagged indices has a ~3% improvement in RSS on `mse`, no change on `serde1`. But I see a ~50% improvement in the runtime of `mse` and a similarly-large regression on `serde1`. Which is itself interesting; some workloads are very Untagged-heavy but don't have an Untagged near the top of the stack.

So I of course slapped a trivial cache on it, and now it's an improvement across the board.

I _was_ planning on using `core` as my benchmarks suite, but of course the `join` test is failing :cry: so I'm going to collect some interesting cases from third-party crates I've run `cargo miri test` on.