@RalfJung 2021-12-20
> I was planning on using core as my benchmarks suite, but of course the join test is failing

https://github.com/rust-lang/rust/pull/92121 should help with that.