@saethlin 2022-01-07
I'm still slowly working on this.

I tried running miri-test-libstd over `core` with `-Zmiri-tag-raw-pointers` and I killed the process after 2 hours (using 20 GB). It spent over an hour doing _something_ after printing
```
Failures:
```
So I have absolutely no idea what was going on there. It would be cool to have a way to peer inside miri and see what code it is executing, best guess is something in the test harness that ends up quadratic or worse inside miri due to the nature of stacked borrows.

I also tried running it the tests for `core` with this branch, and the memory usage started growing rapidly at the same stage in the run, I killed the process when it hit 100 GB. Neither of these runs completed, so I don't think there is a 5x memory blowup from this PR, it seems likely that memory would have only grown to this level if I'd let the current tip of master run long enough.

At the moment I'm chewing on ideas to do something about this. I don't think a patch that makes running the standard library tests with tag-raw-pointers a non-starter is a good idea.