@RalfJung 2022-01-08
> t would be cool to have a way to peer inside miri and see what code it is executing,

Yeah that would be cool... Cc https://github.com/rust-lang/miri/issues/1782.
The issue you describe sounds familiar as well: https://github.com/rust-lang/miri/issues/1780. However it is odd that one of the tests would fail...