@RalfJung 2022-03-05
> I'm still slowly working on this.

Marking as "waiting for author" based on above comment. Please let me know if that is inaccurate. :)